Run and list-types commands
- bd:query:list-types will list QueryTypes, and let you see details for each (service, class, parameters)
- bd:query:run will run a Query from a QueryType

query:run is meant to be extensible. Rendering the SearchResult is done through a SearchResultFormatter.
Changing the output can be done by setting a new service for the command, or creating a new command
with a different SearchResultFormatter.

The SimpleSearchResultFormatter, used by default, delegates rendering of each SearchHit to a SearchHitFormatter.
A default ExpressionLanguage provides a one_line formatter. A new one could be created, with a different expression,
and given to the search_result_formatter.

The list of QueryTypes is collected by a compiler pass, using the arguments from the QueryTypeRegistry.